app.jsにPrometheusメトリクスの初期化処理を移動し、db-dev.jsonに新しいイベントデータを追加しました。また、テスト用HTTPリクエストのイベントIDをEVT-004に更新しました。
