+++ api/database/db-dev.json
@@ -82,6 +82,17 @@
       "part_hash": "a3c723421d92af55c1bbc1fa8b82b7a0268a4c2147649b7c9f7bb68c04f8b164",
       "supplier_hash": "713fd982a8a65d0c28c95a4d8ff48780fb18f5349c0627f7c956cff0a986d13b",
       "operator_hash": "627b350dcccec4eb28a30ac3ae6cac5c6ba6e47f6a465a802dc053be5ce0d61b"
+    },
+    {
+      "event_id": "EVT-004",
+      "part_id": "BAT-001",
+      "status": "assembled",
+      "timestamp": "2025-03-26T16:00:00Z",
+      "location": "ENEGEN 本社工場",
+      "operator_id": "USR-9001",
+      "part_hash": "a3c723421d92af55c1bbc1fa8b82b7a0268a4c2147649b7c9f7bb68c04f8b164",
+      "supplier_hash": "713fd982a8a65d0c28c95a4d8ff48780fb18f5349c0627f7c956cff0a986d13b",
+      "operator_hash": "627b350dcccec4eb28a30ac3ae6cac5c6ba6e47f6a465a802dc053be5ce0d61b"
     }
   ]
 }
