+++ api/test/api.http
@@ -1,5 +1,5 @@
 # @event_id = EVT-{{$timestamp}}
-@event_id = EVT-003
+@event_id = EVT-004
 
 ### 🔹 オンチェーン登録：POST /api/part-event
 POST http://localhost:8080/api/part-event
